macOS: libc++ string includes in crypto_onwire.h; uint64 encode/decode in metrics Types.h
- msg/async/crypto_onwire.h: add <string>/<string_view> (libc++ does not pull
  them transitively; std::string/string_view used in the header). The cascading
  'only virtual member functions can be marked override' errors were fallout from
  the incomplete std::string type.
- include/cephfs/metrics/Types.h: encode/decode of subvolume_metrics size used
  size_t, which on Darwin arm64 (unsigned long) differs from uint64_t (unsigned
  long long), making the encode<>/decode<> overload set ambiguous. Pin to uint64_t.

+++ src/include/cephfs/metrics/Types.h
@@ -754,7 +754,9 @@ struct SubvolumeMetricsPayload : public ClientMetricPayloadBase {
   void encode(bufferlist &bl) const {
     using ceph::encode;
     ENCODE_START(1, 1, bl);
-    encode(subvolume_metrics.size(), bl);
+    // size() is size_t (unsigned long); on Darwin that differs from uint64_t
+    // (unsigned long long) so the encode<> overload set is ambiguous. Pin it.
+    encode((uint64_t)subvolume_metrics.size(), bl);
     for(auto const& m : subvolume_metrics) {
       m.encode(bl);
     }
@@ -764,7 +766,7 @@ struct SubvolumeMetricsPayload : public ClientMetricPayloadBase {
   void decode(bufferlist::const_iterator &iter) {
     using ceph::decode;
     DECODE_START(1, iter) ;
-    size_t size = 0;
+    uint64_t size = 0;  // decode<> has no size_t overload on Darwin; use uint64_t
     decode(size, iter);
     subvolume_metrics.clear();
     subvolume_metrics.reserve(size);
